Attempting to debug the failing unit test.
Added a new method to AddBooksForm to print out the validation state of every
field. This reports that every field is valid so I'm wondering why the form
validation is not succeeding.

Note that unit test failure says that the return code is 400 and that only happens
if the form did not validate.