At galadriel, the database transactions seem to be working fine.
But of course there are problems.

The first problem is that the request to add books got blocked by login_required.
So I tried to call `login_user` on `_creator`.

Then I got this error "working outside of request context". So I tried to work
within contet as I found at StackOverflow. But still no!