Trying to test get_or_create but having problems.
The problem is in the creation of objects from factories. In the included test,
a `Genre` is created with the creator field submitted as a username (as opposed
to an ID). This is where the docs conflict.

The docs for factory_boy: http://factoryboy.readthedocs.org/en/latest/recipes.html#dependent-objects-foreignkey
and http://factoryboy.readthedocs.org/en/latest/orms.html#sqlalchemy

The id PK is declared among the fields of the factory. However, if I do that in
the present code base, the id will be passed as a parameter to the ORM object
constructors, which causes an error.

The models in this code is patterned after
https://pythonhosted.org/Flask-SQLAlchemy/models.html.