Attempt for #44.
But won't be tested until I have to enter a new contributor more than once.

The problem is that the autocomplete list keeps the filtered lists even after
the book is submitted. Hence even if I'm adding in the names of new contributors
when the submission succeeds, the next entry is unable to benefit from it.